Rename conjugate_if_needed.hpp to conj_if_needed.hpp
This better reflects the header's contents.

Also, fix the kokkos-kernels implementation, as it no longer needs
the is_complex trait to determine if conj should be called
on a number type.
